backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose
pytest
python-multipart
httpx
